requirements.txt
flask==3.0.3
flask-cors==4.0.1
torch==2.3.0
torchvision==0.18.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
